+++ CHANGELOG.md
@@ -39,6 +39,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Improvements
 
+* [#646](https://github.com/babylonlabs-io/finality-provider/pull/646) chore: adds more e2e for cosmos
 * [#647](https://github.com/babylonlabs-io/finality-provider/pull/647) chore: unify methods in eotsmanager
 
 ## v2.0.0-rc.3